+++ crates/rib/src/manager/distribution/bgpls.rs
@@ -250,8 +250,8 @@ impl RibManager {
             if (!bgpls_announce.is_empty() || !bgpls_withdraw.is_empty())
                 && !self.try_send_and_commit_outbound_update(
                     peer,
-                    vec![],
-                    vec![],
+                    vec![].into(),
+                    vec![].into(),
                     vec![],
                     vec![],
                     vec![],