+++ crates/rib/src/manager/distribution/evpn.rs
@@ -462,8 +462,8 @@ impl RibManager {
             if (!evpn_announce.is_empty() || !evpn_withdraw.is_empty())
                 && !self.try_send_and_commit_outbound_update(
                     peer,
-                    vec![],
-                    vec![],
+                    vec![].into(),
+                    vec![].into(),
                     vec![],
                     vec![],
                     vec![],