+++ crates/rib/src/manager/distribution/flowspec.rs
@@ -376,8 +376,8 @@ impl RibManager {
             if (!fs_announce.is_empty() || !fs_withdraw.is_empty())
                 && !self.try_send_and_commit_outbound_update(
                     peer,
-                    vec![],
-                    vec![],
+                    vec![].into(),
+                    vec![].into(),
                     vec![],
                     vec![],
                     vec![],