+++ crates/rib/src/manager/distribution/rtc.rs
@@ -270,8 +270,8 @@ impl RibManager {
             if (!rtc_announce.is_empty() || !rtc_withdraw.is_empty())
                 && !self.try_send_and_commit_outbound_update(
                     peer,
-                    vec![],
-                    vec![],
+                    vec![].into(),
+                    vec![].into(),
                     vec![],
                     vec![],
                     vec![],